Handlers: Properly encode the names of projects and repostories
Projects and repositories which aren't url encoded cause problems
when launching external utilities to view the code reviews. For
instance if your project has a name in it, the URL will be malformed
when round tripped through the command line.

This also fixes cases where we were assuming project name == repo name.
Which is not the case in many projects, besies one of the main ones I
test. :) This makes the tool a whole lot more useful to most people.