DataSource: Fetch reviews from all repositories in specified project
This change cleans up the configuration, and makes the behavior a bit
cleaner with respect to what users expect. Previously we had cray logic
which assumed if you didn't specify the repo name, that we should assume
the project name is the same as repro name. This worked for some of the
repositories I work on, but not the general case.

The Azure DevOps API introduced a new preview API which lets you fetch
all pull requests for all repo's under a single project. Use this API
instead of the previous crazy logic, and delete all the old code as it
is now unused.