PullRequestView: Fix refresh when no reviews were loaded previously
We were previously trying to set the selected index even if there was
no data available, resulting in the following exception:

    System.ArgumentException: value
       at Terminal.Gui.ListView.set_SelectedItem(Int32 value)
       at PrDash.View.PullRequestView.RefreshCallback() in D:\enl\pr-dash\src\View\PullRequestView.cs:line 327

Fix this, and improve loading expereince by re-setting loading state
on load, and then we will always have something to select. This also
cleans up the force refresh when no rows are available case.

Reported-By: @hans-olav