Views: Fix potential race in PullRequestView initialization
The first render of the view was being started before we had added the view to
the UI tree, so it seemed like it was technically possible to start trying to
render before we were in the tree. This  may or may not result in bugs on the
initial rendering. To avoid this just cleanup the setup to we only start the
population/refresh task after everything is setup.